+++ R/daemons.R
@@ -425,11 +425,18 @@ status <- function(.compute = "default") {
 #'     receiving external pointer reference objects.
 #'
 #' @param refhook \strong{either} a list or pairlist of two functions: the
-#'     signature for the first must accept a list of external pointer type
-#'     objects and return a raw vector, e.g. \code{torch::torch_serialize}, and
-#'     the second must accept a raw vector and return a list of external pointer
-#'     type objects, e.g. \code{torch::torch_load},\cr \strong{or else} NULL to
-#'     reset.
+#'     signature for the first must accept a reference object inheriting from
+#'     'class' (or a list of such objects) and return a raw vector, and the
+#'     second must accept a raw vector and return reference objects (or a list
+#'     of such objects), \cr \strong{or else} NULL to reset.
+#' @param class [default ""] a character string representing the class of object
+#'     that these serialization function will be applied to, e.g. 'ArrowTabular'
+#'     or 'torch_tensor'.
+#' @param list [default FALSE] the serialization functions accept and return
+#'     reference object individually e.g. \code{arrow::write_to_raw} and
+#'     \code{arrow::read_ipc_stream}. If TRUE, the serialization functions
+#'     accept and return a list of reference objects, e.g.
+#'     \code{torch::torch_serialize} and \code{torch::torch_load}.
 #'
 #' @return Invisibly, the pairlist of currently-registered 'refhook' functions.
 #'     A message is printed to the console when functions are successfully
@@ -451,10 +458,10 @@ status <- function(.compute = "default") {
 #'
 #' @export
 #'
-serialization <- function(refhook = list()) {
+serialization <- function(refhook = list(), class = "", list = FALSE) {
 
   register <- !missing(refhook)
-  cfg <- next_config(refhook = refhook)
+  cfg <- next_config(refhook = refhook, class = class, list = list)
 
   if (register) {
     if (is.list(refhook) && length(refhook) == 2L && is.function(refhook[[1L]]) && is.function(refhook[[2L]]))
@@ -462,7 +469,8 @@ serialization <- function(refhook = list()) {
         if (is.null(refhook))
           cat("mirai serialization functions cancelled\n", file = stderr()) else
             stop(._[["refhook_invalid"]])
-    register_everywhere(refhook)
+    `[[<-`(., "refhook", list(refhook, class, list))
+    register_everywhere(refhook = refhook, class = class, list = list)
   }
 
   invisible(cfg)
@@ -583,11 +591,13 @@ query_status <- function(envir) {
                            dimnames = list(envir[["urls"]], c("i", "online", "instance", "assigned", "complete"))))
 }
 
-register_everywhere <- function(refhook)
+register_everywhere <- function(refhook, class, list)
   for (.compute in names(..))
-    everywhere(mirai::serialization(refhook), refhook = refhook, .compute = .compute)
+    everywhere(mirai::serialization(refhook = refhook, class = class, list = list),
+               refhook = refhook, class = class, list = list, .compute = .compute)
 
-serialization_refhook <- function(refhook = next_config())
-  if (length(refhook[[1L]])) register_everywhere(refhook)
+serialization_refhook <- function(refhook = .[["refhook"]])
+  if (length(refhook[[1L]]))
+    register_everywhere(refhook = refhook[[1L]], class = refhook[[2L]], list = refhook[[3L]])
 
 ._scm_. <- as.raw(c(0x07, 0x00, 0x00, 0x00, 0x42, 0x0a, 0x03, 0x00, 0x00, 0x00, 0x02, 0x03, 0x04, 0x00, 0x00, 0x05, 0x03, 0x00, 0x05, 0x00, 0x00, 0x00, 0x55, 0x54, 0x46, 0x2d, 0x38, 0xfc, 0x00, 0x00, 0x00))